v0.5

+++ Plane.h
@@ -9,7 +9,7 @@ public:
 	Plane(glm::vec3 pos, glm::vec3 normal, glm::vec3 amb_col, glm::vec3 diff_col, glm::vec3 spe_col, float shininess);
 	~Plane();
 	bool hit(const Ray& ray, double& t, ShadeInfo& shadeInfo);
-	glm::vec3 pos, normal;
+	glm::dvec3 pos, normal;
 	void print();
 };
 #endif
