+++ Ray.h
@@ -6,11 +6,11 @@ class Ray
 {
 	public:
 		Ray() {}
-		Ray(const glm::vec3& origin, const glm::vec3& dir) { this->origin = origin; this->dir = dir; }
+		Ray(const glm::dvec3& origin, const glm::dvec3& dir) { this->origin = origin; this->dir = dir; }
 		~Ray() {};
-		glm::vec3 getOrigin() const { return this->origin; }
-		glm::vec3 getDirection() const { return this->dir; }
-		glm::vec3 pointAtParameter(float t) const { return this->origin + (t*this->dir); }
-		glm::vec3 origin, dir;
+		glm::dvec3 getOrigin() const { return this->origin; }
+		glm::dvec3 getDirection() const { return this->dir; }
+		glm::dvec3 pointAtParameter(double t) const { return this->origin + (t*this->dir); }
+		glm::dvec3 origin, dir;
 };
 #endif
